refactor(picker): consolidate spawn_precompute ordering onto PreviewOrchestrator
Move the WHAT-and-in-WHAT-ORDER logic from PickerHandler into a new
PreviewOrchestrator::spawn_all_precompute method. The handler now makes a
single delegate call from on_skeleton instead of carrying mode order,
first-item prioritization, and summary gating in its own impl.

This drops the PickerHandler::repo field (only the orchestrator still
needs it) and shrinks the handler's surface to skim-channel glue plus
shared per-row state, leaving spawn ordering owned by the same struct
that owns the rayon pool and cache.

Behavior is unchanged: same mode order, same first-item priority, same
summary-hint fallback when LLM is disabled.
